GaussPattern + SpikeGenerator. Missing unit tests and documentation.
