--- tests/lava/lib/dnf/unit/inputs/bias_source/test_models.py
@@ -1,9 +1,0 @@
-# Copyright (C) 2021 Intel Corporation
-# SPDX-License-Identifier: BSD-3-Clause
-# See: https://spdx.org/licenses/
-
-import unittest
-
-
-if __name__ == '__main__':
-    unittest.main()
