+++ tests/lava/lib/dnf/unit/inputs/patterns/gauss_pattern/test_process.py
@@ -1,0 +1,13 @@
+# Copyright (C) 2021 Intel Corporation
+# SPDX-License-Identifier: BSD-3-Clause
+# See: https://spdx.org/licenses/
+
+import unittest
+
+
+class TestGaussPattern(unittest.TestCase):
+    pass
+
+
+if __name__ == '__main__':
+    unittest.main()
